+++ src/ConfigEditor.tsx
@@ -1,13 +1,13 @@
 import React, { useCallback } from 'react';
 import { DataSourcePluginOptionsEditorProps, DataSourceSettings } from '@grafana/data';
-import { DataSourceOptions } from './types';
-import { FieldSet, Form, InlineField, Input } from '@grafana/ui';
+import { DataSourceOptions, DataSourceSecureOptions } from './types';
+import { FieldSet, Form, InlineField, Input, SecretInput } from '@grafana/ui';
 
-interface Props extends DataSourcePluginOptionsEditorProps<DataSourceOptions> {}
+interface Props extends DataSourcePluginOptionsEditorProps<DataSourceOptions, DataSourceSecureOptions> { }
 
 export const ConfigEditor = ({ options, onOptionsChange }: Props): JSX.Element => {
   const onSettingsChange = useCallback(
-    (change: Partial<DataSourceSettings<DataSourceOptions>>) => {
+    (change: Partial<DataSourceSettings<DataSourceOptions, DataSourceSecureOptions>>) => {
       onOptionsChange({
         ...options,
         ...change,
@@ -49,6 +49,17 @@ export const ConfigEditor = ({ options, onOptionsChange }: Props): JSX.Element =
                   placeholder="materialize"
                   onChange={(event) =>
                     onSettingsChange({ jsonData: { ...options.jsonData, username: event.currentTarget.value } })
+                  }
+                />
+              </InlineField>
+
+              <InlineField label="Password" labelWidth={20}>
+                <SecretInput
+                  isConfigured={!!options.secureJsonData?.password ?? false}
+                  onReset={() => onSettingsChange({ secureJsonData: { ...options.secureJsonData, password: undefined } })}
+                  value={options.secureJsonData?.password}
+                  onBlur={(event) =>
+                    onSettingsChange({ secureJsonData: { ...options.secureJsonData, password: event.currentTarget.value } })
                   }
                 />
               </InlineField>